# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KYC API",
    "version": "1.0.0",
    "description": "JSON:API endpoints for Applicant accounts and sessions, KYC applications and document evidence, reviewer workload, and access checks. Authenticated endpoints use the first-party session cookie. Error responses use a JSON:API errors array with safe, client-actionable details."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current frontend origin; /api/v1 is proxied to the Java API in development."
    }
  ],
...
  "paths": {
...
    "/api/v1/applicant-applications/{applicationId}/form": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        }
      ],
      "get": {
        "operationId": "getApplicantApplicationForm",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authenticated Applicant's form state.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "summary": "Get an Applicant application form",
        "description": "Returns the authenticated Applicant?s saved draft form, including current step, step progress, saved answers, document-evidence presence, and form version. The application must belong to the signed-in Applicant."
      },
      "patch": {
        "operationId": "updateApplicantApplicationForm",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/FormUpdateDocument"
              }
            }
          },
          "description": "JSON:API document containing the application form resource, selected step, answers for that step, and optional expected form version."
        },
        "responses": {
          "200": {
            "description": "The persisted form state after the update.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDocument"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/MalformedJsonApiDocument"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "$ref": "#/components/responses/FormVersionConflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedRequestMediaType"
          },
          "422": {
            "$ref": "#/components/responses/FormUnprocessable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "required": false,
            "description": "CSRF token from the XSRF-TOKEN cookie. Required when sending a session cookie with a state-changing request.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Save Applicant application form answers",
        "description": "Saves answers for the selected form step and returns the persisted form state. Include the current form version to detect concurrent changes. Only fields for the selected step are accepted; stale versions return 409, invalid form data returns 422, and authenticated writes require the CSRF token."
      }
    },
    "/api/v1/applicant-applications/{applicationId}/review": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        }
      ],
      "get": {
        "operationId": "getApplicantApplicationReview",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Missing required fields or a ready-to-submit review summary.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "summary": "Get the Applicant application readiness review",
        "description": "Returns whether the draft is ready for submission and lists missing required fields by step and field name. This response does not submit or change the application."
      }
    },
    "/api/v1/applicant-applications/{applicationId}/document-evidence": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        }
      ],
      "post": {
        "operationId": "uploadApplicantApplicationDocumentEvidence",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Identity document image. Use JPG (image/jpeg) or PNG (image/png); maximum size is 10 MiB."
                  }
                }
              }
            }
          },
          "description": "Multipart upload with one file part. Accepts image/jpeg or image/png files up to 10 MiB."
        },
        "responses": {
          "201": {
            "description": "Document evidence stored for the Applicant's draft.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentEvidenceDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedEvidenceRequestMediaType"
          },
          "422": {
            "$ref": "#/components/responses/EvidenceInvalid"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "required": false,
            "description": "CSRF token from the XSRF-TOKEN cookie. Required when sending a session cookie with a state-changing request.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Upload Applicant identity document evidence",
        "description": "Stores one JPG or PNG image for the authenticated Applicant?s draft. Send multipart/form-data with a file part no larger than 10 MiB. An empty, oversized, or unsupported image returns a 422 JSON:API error; authenticated uploads require the CSRF token."
      }
    },
    "/api/v1/applicant-applications/{applicationId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        }
      ],
      "get": {
        "operationId": "getApplicantApplicationStatus",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authenticated Applicant's minimized application lifecycle state.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "summary": "Get Applicant application status",
        "description": "Returns the authenticated Applicant?s minimized application lifecycle state. The resource omits personal form answers and document contents."
      },
      "patch": {
        "operationId": "submitApplicantApplication",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequestDocument"
              }
            }
          },
          "description": "JSON:API document confirming submission and carrying the expected form version returned by the latest form read."
        },
        "responses": {
          "200": {
            "description": "Submission committed or the same already-committed submission returned after a replay.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDocument"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/MalformedJsonApiDocument"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "$ref": "#/components/responses/FormVersionConflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedRequestMediaType"
          },
          "422": {
            "$ref": "#/components/responses/SubmissionUnprocessable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "required": false,
            "description": "CSRF token from the XSRF-TOKEN cookie. Required when sending a session cookie with a state-changing request.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Submit the Applicant application",
        "description": "Submits a complete draft after explicit confirmation and an expected form version. A successful replay returns the already-committed submission. Incomplete applications return 422 with missing field names in the error detail; stale versions or invalid lifecycle state return 409."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-KYCSESSION",
        "description": "Opaque, secure, HttpOnly cookie set after successful sign-in. The browser sends it automatically to the API origin."
      }
    },
    "schemas": {
...
      "FormDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FormResource"
          }
        },
        "description": "JSON:API response containing the saved form state for one application."
      },
...
      "FormUpdateDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["type", "id", "attributes"],
            "properties": {
              "type": {
                "type": "string",
                "const": "applicant-application-forms"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "attributes": {
                "$ref": "#/components/schemas/FormUpdateAttributes"
              }
            }
          }
        },
        "description": "JSON:API request document for saving answers on one form step."
      },
...
    },
    "responses": {
...
      "AuthenticationRequired": {
        "description": "No active Applicant session is available. Sign in and send the session cookie.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "authenticationRequired": {
                "summary": "Session required",
                "value": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "authentication-required",
                      "title": "Authentication required",
                      "detail": "Sign in to continue."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "AccessDenied": {
        "description": "The request is not authorized for this resource or is missing a valid CSRF token. The response does not disclose protected resource details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "accessDenied": {
                "summary": "Access denied",
                "value": {
                  "errors": [
                    {
                      "status": "403",
                      "code": "access-denied",
                      "title": "Access denied",
                      "detail": "You do not have permission to access this resource."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "UnsupportedRequestMediaType": {
        "description": "The request Content-Type is not accepted for this operation. JSON:API request bodies must use application/vnd.api+json. For document-evidence uploads, use multipart/form-data; image validation failures return 422.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "unsupportedMediaType": {
                "summary": "Unsupported request media type",
                "value": {
                  "errors": [
                    {
                      "status": "415",
                      "code": "unsupported-media-type",
                      "title": "Unsupported media type",
                      "detail": "Use application/vnd.api+json for JSON:API request bodies."
                    }
                  ]
                }
              },
              "mediaTypeParameters": {
                "summary": "Unsupported JSON:API media type parameters",
                "value": {
                  "errors": [
                    {
                      "status": "415",
                      "code": "unsupported-media-type",
                      "title": "Unsupported media type",
                      "detail": "Use application/vnd.api+json without media type parameters."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "NotAcceptable": {
        "description": "The requested response representation is not supported. Accept application/vnd.api+json.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "notAcceptable": {
                "summary": "Unsupported response representation",
                "value": {
                  "errors": [
                    {
                      "status": "406",
                      "code": "not-acceptable",
                      "title": "Not acceptable",
                      "detail": "Accept application/vnd.api+json for JSON:API responses."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "MalformedJsonApiDocument": {
        "description": "The JSON request body could not be parsed as a JSON:API document.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "malformedDocument": {
                "summary": "Malformed JSON:API document",
                "value": {
                  "errors": [
                    {
                      "status": "400",
                      "code": "malformed-document",
                      "title": "Malformed JSON:API document",
                      "detail": "The request body is not a valid JSON:API document."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "ApplicationNotFound": {
        "description": "The application is absent or is not visible to the authenticated Applicant.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "applicationNotFound": {
                "summary": "Application not found",
                "value": {
                  "errors": [
                    {
                      "status": "404",
                      "code": "application-not-found",
                      "title": "Application not found",
                      "detail": "The requested application is not available."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "FormVersionConflict": {
        "description": "The form version changed or the application cannot be submitted in its current state. Reload the form before retrying.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "formVersionConflict": {
                "summary": "Form version or lifecycle conflict",
                "value": {
                  "errors": [
                    {
                      "status": "409",
                      "code": "form-version-conflict",
                      "title": "Form changed",
                      "detail": "Reload the form and try again."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The request could not be completed. The response omits internal exception and persistence details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "internalError": {
                "summary": "Unexpected server error",
                "value": {
                  "errors": [
                    {
                      "status": "500",
                      "code": "internal-error",
                      "title": "Request failed",
                      "detail": "The request could not be completed."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The persistence operation failed. The response uses the safe fixed message shown below and omits database details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "submissionUnavailable": {
                "summary": "Persistence unavailable",
                "value": {
                  "errors": [
                    {
                      "status": "503",
                      "code": "submission-unavailable",
                      "title": "Submission unavailable",
                      "detail": "Your application was not submitted. Please retry later."
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "ApplicationId": {
        "name": "applicationId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "UUID of the application to read or update. The application must be visible to the authenticated Applicant."
      }
    },
...
  }
}